fix issues

+++ src/app/flows/flows.component.ts
@@ -201,14 +201,16 @@ export class FlowsComponent implements OnInit {
     const selectedFIIds = this.selection.selected.map(row => {
       return { ID: row.ID};
     });
-    for (let fi of selectedFIIds) {
-      this.tresService.deleteFi(fi.ID).subscribe(
-        (data: INd) => {
-          this.initDataSource(data);
-        },
-        error => {
-          console.error(error);
-        });
+    if (confirm('Are you sure?')) {
+      for (let fi of selectedFIIds) {
+        this.tresService.deleteFi(fi.ID).subscribe(
+          (data: INd) => {
+            this.initDataSource(data);
+          },
+          error => {
+            console.error(error);
+          });
+      }
     }
   }
 }
